+++ src/settings.rs
@@ -1,5 +1,5 @@
 use crate::{cache::CacheMode, json};
-use anyhow::{anyhow, Context};
+use anyhow::Context;
 use log::debug;
 use reqwest::Url;
 use serde::{Deserialize, Serialize};
@@ -105,21 +105,13 @@ impl Settings {
     pub fn set_source(&mut self, source: Source) -> Result<(), anyhow::Error> {
         debug!("Set source to {}", source.path.display());
 
-        let source_paths = self
-            .sources
-            .iter()
-            .map(|source| &source.path)
-            .collect::<Vec<_>>();
-
-        if !source_paths.contains(&&source.path) {
-            self.sources.push(source);
-            Ok(())
+        if let Some(s) = self.sources.iter_mut().find(|s| s.path == source.path) {
+            *s = source;
         } else {
-            Err(anyhow!(format!(
-                "Source already configured: {}",
-                source.path.display()
-            )))
+            self.sources.push(source);
         }
+
+        Ok(())
     }
 
     pub fn set_cache_mode(&mut self, cache_mode: Option<CacheMode>) {
@@ -191,12 +183,11 @@ mod tests {
     #[test]
     fn test_set_source() {
         let mut settings = Settings::default();
-        settings
-            .set_source(Source {
-                path: PathBuf::from("path/to/source"),
-                folders: vec![],
-            })
-            .unwrap();
+        let res = settings.set_source(Source {
+            path: PathBuf::from("path/to/source"),
+            folders: vec![],
+        });
+        assert!(res.is_ok());
         assert_eq!(
             settings,
             Settings {
@@ -210,7 +201,7 @@ mod tests {
     }
 
     #[test]
-    fn test_set_source_duplicate() {
+    fn test_set_source_overwrite() {
         let mut settings = Settings::default();
         settings
             .set_source(Source {
@@ -229,11 +220,22 @@ mod tests {
             }
         );
 
-        let res = settings.set_source(Source {
-            path: PathBuf::from("path/to/source"),
-            folders: vec![],
-        });
-        assert!(res.is_err());
+        settings
+            .set_source(Source {
+                path: PathBuf::from("path/to/source"),
+                folders: vec!["dev".to_string(), "articles".to_string()],
+            })
+            .unwrap();
+        assert_eq!(
+            settings,
+            Settings {
+                sources: vec![Source {
+                    path: PathBuf::from("path/to/source"),
+                    folders: vec!["dev".to_string(), "articles".to_string()]
+                }],
+                ..Default::default()
+            }
+        );
     }
 
     #[test]